dts: exynos 9810: Reduce charge voltage earlier
Switch to a lower charge voltage more quickly (after 20 cycles, rather than 300 stock). This causes a slightly reduce battery capacity as the -20mV results in about 3% less usable capacity but this helps to extend the cycle life of the battery which is why we want to do this sooner rather than later.